Created a more descriptive README file.
I came to this repository to add to this app using my knowledge in Flutter, but as the app is not nearly completed yet and I cannot figure out what it should finally look like, I decided to create a better README file to describe the project. Please pull it if you find it useful.